requirements.txt
annotated-types==0.7.0
base58==2.1.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.4.1
fonttools==4.55.3
idna==3.10
importlib_metadata==8.5.0
kiwisolver==1.4.8
loguru==0.6.0
matplotlib==3.10.0
mplfinance==0.12.10b0
msgpack==1.1.0
numpy==2.2.1
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyNaCl==1.5.0
pyparsing==3.2.1
pyrsca==0.1.2
pysolace==0.9.40
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
sentry-sdk==1.5.12
setuptools==75.1.0
shioaji==1.2.5
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wheel==0.44.0
xxhash==3.5.0
zipp==3.21.0
